refactor: cenrtalised metrics
