--- .env.local
@@ -1,4 +1,0 @@
-NODE_VERSION=
-PROMETHEUS_VERSION=v3.4.0
-PROMETHEUS_PUSHGATEWAY_VERSION=v1.11.1
-GRAFANA_VERSION=12.0.1
