+++ .github/workflows/checks.yml
@@ -5,8 +5,6 @@ on:
     branches-ignore:
       - main
       - release
-    paths:
-      - '.github/workflows/checks.yml'
   workflow_call:
     inputs:
       scan:
