+++ apps/admin/src/app/api/status/route.ts
@@ -1,3 +1,3 @@
 import handler from '@shared/routes/status';
 
-export const GET = () => handler();
+export const GET = handler();
