+++ apps/app/public/styles/variables.css
@@ -1,0 +1,1 @@
+/* This is a dummy file to satisfy the import of variables.css in the app. */
