+++ apps/app/src/app/api/status/route.ts
@@ -1,4 +1,3 @@
-import { withMetrics } from '@app/utilities/metrics';
 import handler from '@shared/routes/status';
 
-export const GET = withMetrics(handler, '/status');
+export const GET = handler();
