+++ development/grafana/slack-contacts-point.json
@@ -3,7 +3,7 @@
   "type": "slack",
   "uid": "slack-alerts",
   "settings": {
-      "token": "overrideme",
+    "token": "overrideme",
     "recipient": "#alerting-dev"
   },
   "secureSettings": {
