+++ development/prometheus/prometheus.yml
@@ -3,6 +3,17 @@ global:
   scrape_timeout: 10s
 
 scrape_configs:
+  - job_name: adequate_admin_metrics
+    honor_timestamps: true
+    metrics_path: /metrics
+    scheme: http
+    follow_redirects: true
+    static_configs:
+      - targets:
+          - admin:3000 # Domain of your website
+        labels:
+          project: 'adequate'
+          service: 'admin'
   - job_name: adequate_app_metrics
     honor_timestamps: true
     metrics_path: /metrics
@@ -12,6 +23,7 @@ scrape_configs:
       - targets:
           - app:3000 # Domain of your website
         labels:
+          project: 'adequate'
           service: 'app'
   - job_name: 'pushgateway'
     static_configs:
