+++ packages/shared/package.json
@@ -4,6 +4,7 @@
   "private": true,
   "dependencies": {
     "@upstash/ratelimit": "^2.0.5",
-    "@upstash/redis": "^1.35.0"
+    "@upstash/redis": "^1.35.0",
+    "prom-client": "15.1.3"
   }
 }
